Add registration, logout, and authorisation checks.

+++ app/Http/Controllers/ChirpController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Models\Chirp;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -37,10 +40,7 @@ class ChirpController extends Controller
             'message.max' => 'You only have 255 characters.'
         ]);
 
-        Chirp::create([
-            'message' => $validated['message'],
-            // 'user_id' => whatever
-        ]);
+        auth()->user()->chirps()->create($validated);
 
         return redirect('/')->with('success', 'Your Chirp was posted');
     }
@@ -66,6 +66,8 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         $validated = $request->validate([
             'message' => 'required|string|max:255'
         ], [
@@ -84,7 +86,7 @@ class ChirpController extends Controller
      */
     public function destroy(Chirp $chirp)
     {
-        // $this->authorize('delete', $chirp);
+        $this->authorize('delete', $chirp);
 
         $chirp->delete();
